updated code

+++ src/remove_activations.py
@@ -20,7 +20,7 @@ def main(instance_path: str, start_index: int, end_index: int):
         _, dirnames, _ = next(os.walk(os.getcwd(), ))
         for seed_dir in dirnames:
             for fname in os.listdir(Path(os.getcwd(), seed_dir)):
-                if fname != "checkpoint.pt":
+                if fname != "checkpoint.pt" and '.pt' in fname:
                     os.remove(Path(Path(os.getcwd(), seed_dir), fname))
 
         os.chdir('..')
